requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
Faker==18.4.0
fastapi==0.95.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pydantic==1.10.7
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
